+++ src/ro/ase/mpai/platiStrategy/ShoppingCart.java
@@ -2,7 +2,7 @@ package ro.ase.mpai.platiStrategy;
 
 public class ShoppingCart {
 
-	public void plata(IPlati metodaPlata, int suma) {
+	public void plata(IPlati metodaPlata, double suma) {
 		metodaPlata.plata(suma);
 	}
 }
